Applied changes based on rust styleguide
